Move the tests from the removed exception class and remove the file
