+++ tests/unit/UncountableValueExceptionTest.php
@@ -16,28 +16,94 @@ use MAChitgarha\Exception\JSON\UncountableValueException;
 
 class UncountableValueExceptionTest extends TestCase
 {
-    /** @var JSON */
-    protected $json;
-
     protected function setUp()
     {
         $this->expectException(UncountableValueException::class);
+    }
 
-        $this->json = new JSON(file_get_contents(__DIR__ . "/../json/apps.json"));
+    protected function prependJsonToEveryElement(JSON $json, array $array): array
+    {
+        foreach ($array as &$item) {
+            array_unshift($item, $json);
+        }
+        return $array;
     }
 
     /** @dataProvider methodsWithUncountableIndexProvider */
-    public function testCallMethodsOnUncountableIndex(string $funcName, ...$params)
+    public function testCallMethodsOnUncountableIndex(JSON $json, string $funcName, ...$args)
     {
-        $this->json->$funcName(...$params);
+        $json->$funcName(...$args);
+    }
+
+    /** @dataProvider callMethodsWithIndexesProvider */
+    public function testCallMethodsWithIndexes(JSON $json, string $funcName, array $args)
+    {
+        $json->$funcName(...$args);
+    }
+
+    /** @dataProvider methodsForCountableDataProvider */
+    public function testMethodsDependOnCountableData(JSON $json, string $funcName, array $args = [])
+    {
+        $json->$funcName(...$args);
+    }
+
+    public function callMethodsWithIndexesProvider()
+    {
+        $json = clone new JSON("Ubuntu");
+        $sampleIndex = "appName";
+        $sampleValue = "Terminal";
+
+        return $this->prependJsonToEveryElement($json, [
+            ["get", [$sampleIndex]],
+            ["set", [$sampleValue, $sampleIndex]],
+            ["unset", [$sampleIndex]],
+            ["isSet", [$sampleIndex]],
+            ["isCountable", [$sampleIndex]]
+        ]);
+    }
+
+    public function methodsForCountableDataProvider()
+    {
+        $json = clone new JSON(1400);
+        $sampleValue = 1398;
+
+        return $this->prependJsonToEveryElement($json, [
+            ["count"],
+            ["iterate"],
+            ["forEach", [function () {}]],
+            ["forEachRecursive", [function () {}]],
+            ["push", [$sampleValue]],
+            ["pop"],
+            ["shift"],
+            ["unshift", [$sampleValue]],
+            ["getValues"],
+            ["getKeys"],
+            ["getRandomValue"],
+            ["getRandomKey"],
+            ["getRandomElement"],
+            ["getRandomValues", [2]],
+            ["getRandomKeys", [2]],
+            ["getRandomSubset", [2]],
+            ["mergeWith", [[]]],
+            ["mergeRecursivelyWith", [[]]],
+            ["difference", [[]]],
+            ["filter"],
+            ["flipValuesAndKeys"],
+            ["reduce", [function () {}]],
+            ["shuffle"],
+            ["reverse"],
+            ["fill", [0, 1, $sampleValue]]
+        ]);
     }
 
     public function methodsWithUncountableIndexProvider()
     {
-        return [
+        $json = new JSON(file_get_contents(__DIR__ . "/../json/apps.json"));
+
+        return $this->prependJsonToEveryElement($json, [
             ["count", "apps.others.0"],
             ["pop", "apps.ides.Microsoft.1"],
             ["shift", "apps.others.2.0"],
-        ];
+        ]);
     }
 }
